Moving the test method into SerializerTests
