+++ src/NServiceBus.Core.Tests/NServiceBus.Core.Tests.csproj
@@ -170,7 +170,6 @@
     <Compile Include="Sagas\SagaMetadataCreationTests.cs" />
     <Compile Include="Recoverability\SecondLevelRetries\DefaultRetryPolicyTests.cs" />
     <Compile Include="Serializers\SerializeMessagesBehaviorTests.cs" />
-    <Compile Include="Serializers\XML\Issue_2796.cs" />
     <Compile Include="StandardsTests.cs" />
     <Compile Include="Encryption\ConfigureRijndaelEncryptionServiceTests.cs" />
     <Compile Include="Encryption\RijndaelEncryptionServiceTest.cs" />
